+++ include/Ark/VM/Value.hpp
@@ -25,7 +25,6 @@
 #include <Ark/VM/Value/Closure.hpp>
 #include <Ark/VM/Value/UserType.hpp>
 #include <Ark/Platform.hpp>
-#include <Ark/Profiling.hpp>
 
 namespace Ark
 {