+++ src/arkreactor/Compiler/Compiler.cpp
@@ -5,6 +5,8 @@
 #include <limits>
 #include <filesystem>
 #include <picosha2.h>
+#include <termcolor/termcolor.hpp>
+#undef max
 
 #include <Ark/Literals.hpp>
 #include <Ark/Utils.hpp>
@@ -260,6 +262,12 @@ namespace Ark
         throw CompilationError(makeNodeBasedErrorCtx(message, node));
     }
 
+    void Compiler::compilerWarning(const std::string& message, const Node& node)
+    {
+        if (m_options & FeatureShowWarnings)
+            std::cerr << termcolor::yellow << "Warning " << termcolor::reset << makeNodeBasedErrorCtx(message, node) << "\n";
+    }
+
     void Compiler::compileExpression(const Node& x, int p, bool is_result_unused, bool is_terminal, const std::string& var_name)
     {
         // register symbols
@@ -364,7 +372,10 @@ namespace Ark
             page(p).emplace_back(Instruction::LOAD_SYMBOL, addSymbol(x));  // using the variable
 
         if (is_result_unused)
+        {
+            compilerWarning("Statement has no effect", x);
             page(p).push_back(Instruction::POP);
+        }
     }
 
     void Compiler::compileSpecific(const Node& c0, const Node& x, int p, bool is_result_unused)
@@ -397,8 +408,11 @@ namespace Ark
         // put inst and number of arguments
         page(p).emplace_back(inst, computeSpecificInstArgc(inst, argc));
 
-        if (is_result_unused && name != "pop!")  // pop! never pushes a value
+        if (is_result_unused && name.back() != '!')  // in-place functions never push a value
+        {
+            compilerWarning("Ignoring return value of function", x);
             page(p).push_back(Instruction::POP);
+        }
     }
 
     void Compiler::compileIf(const Node& x, int p, bool is_result_unused, bool is_terminal, const std::string& var_name)
@@ -469,7 +483,10 @@ namespace Ark
 
         // if the computed function is unused, pop it
         if (is_result_unused)
+        {
+            compilerWarning("Unused declared function", x);
             page(p).push_back(Instruction::POP);
+        }
     }
 
     void Compiler::compileLetMutSet(Keyword n, const Node& x, int p)
@@ -524,7 +541,10 @@ namespace Ark
         page(p).emplace_back(Instruction::LOAD_CONST, id);
 
         if (is_result_unused)
+        {
+            compilerWarning("Unused quote expression", x);
             page(p).push_back(Instruction::POP);
+        }
     }
 
     void Compiler::compilePluginImport(const Node& x, int p)