+++ src/arkreactor/VM/VM.cpp
@@ -1,11 +1,9 @@
-#define NOMINMAX
-
 #include <Ark/VM/VM.hpp>
 
 #include <numeric>
 #include <limits>
 
-#include <termcolor/termcolor.hpp>
+#include <termcolor/proxy.hpp>
 #include <Ark/Files.hpp>
 #include <Ark/Utils.hpp>
 #include <Ark/TypeChecker.hpp>
@@ -301,7 +299,6 @@ namespace Ark
                         else
                             throwVMError("unbound variable: " + m_state.m_symbols[context.last_symbol]);
 
-                        COZ_PROGRESS_NAMED("ark vm load_symbol");
                         break;
                     }
 
@@ -324,7 +321,6 @@ namespace Ark
                             push(&(m_state.m_constants[arg]), context);
                         }
 
-                        COZ_PROGRESS_NAMED("ark vm load_const");
                         break;
                     }
 
@@ -360,9 +356,8 @@ namespace Ark
                             break;
                         }
 
-                        COZ_PROGRESS_NAMED("ark vm store");
 
-                        throwVMError("unbound variable " + m_state.m_symbols[arg] + ", can not change its value");
+                        throwVMError("unbound variable " + m_state.m_symbols[id] + ", can not change its value");
                         break;
                     }
 
@@ -382,7 +377,6 @@ namespace Ark
                         val.setConst(true);
                         (*context.locals.back()).push_back(arg, val);
 
-                        COZ_PROGRESS_NAMED("ark vm let");
                         break;
                     }
 
@@ -446,7 +440,6 @@ namespace Ark
                             push(std::move(ip_or_val), context);
                         }
 
-                        COZ_PROGRESS_NAMED("ark vm ret");
                         break;
                     }
 
@@ -474,7 +467,6 @@ namespace Ark
                         ptr = ptr->valueType() == ValueType::Reference ? ptr->reference() : ptr;
                         (*context.saved_scope.value()).push_back(arg, *ptr);
 
-                        COZ_PROGRESS_NAMED("ark vm capture");
                         break;
                     }
 
@@ -485,9 +477,8 @@ namespace Ark
                             Job: Push the builtin function object on the stack
                         */
 
-                        push(Builtins::builtins[arg].second, context);
 
-                        COZ_PROGRESS_NAMED("ark vm builtin");
+                        push(Builtins::builtins[id].second, context);
                         break;
                     }
 
@@ -509,7 +500,6 @@ namespace Ark
                         else
                             *local = val;
 
-                        COZ_PROGRESS_NAMED("ark vm mut");
                         break;
                     }
 
@@ -528,9 +518,7 @@ namespace Ark
                             break;
                         }
 
-                        COZ_PROGRESS_NAMED("ark vm del");
-
-                        throwVMError("unbound variable: " + m_state.m_symbols[arg]);
+                        throwVMError("unbound variable: " + m_state.m_symbols[id]);
                         break;
                     }
 
@@ -541,8 +529,6 @@ namespace Ark
                             Job: Save the current environment, useful for quoted code
                         */
                         context.saved_scope = context.locals.back();
-
-                        COZ_PROGRESS_NAMED("ark vm save_scope");
                         break;
                     }
 
@@ -585,8 +571,6 @@ namespace Ark
                         */
 
                         loadPlugin(arg, context);
-
-                        COZ_PROGRESS_NAMED("ark vm plugin");
                         break;
                     }
 
@@ -604,8 +588,6 @@ namespace Ark
                         for (uint16_t i = 0; i < arg; ++i)
                             l.push_back(*popAndResolveAsPtr(context));
                         push(std::move(l), context);
-
-                        COZ_PROGRESS_NAMED("ark vm list");
                         break;
                     }
 
@@ -626,8 +608,6 @@ namespace Ark
                         for (uint16_t i = 0; i < arg; ++i)
                             obj.push_back(*popAndResolveAsPtr(context));
                         push(std::move(obj), context);
-
-                        COZ_PROGRESS_NAMED("ark vm append");
                         break;
                     }
 
@@ -656,8 +636,6 @@ namespace Ark
                                 obj.push_back(*it);
                         }
                         push(std::move(obj), context);
-
-                        COZ_PROGRESS_NAMED("ark vm concat");
                         break;
                     }
 
@@ -676,9 +654,6 @@ namespace Ark
                         for (uint16_t i = 0; i < arg; ++i)
                             list->push_back(*popAndResolveAsPtr(context));
 
-                        push(Nil, context);
-
-                        COZ_PROGRESS_NAMED("ark vm append!");
                         break;
                     }
 
@@ -708,9 +683,6 @@ namespace Ark
                                 list->push_back(*it);
                         }
 
-                        push(Nil, context);
-
-                        COZ_PROGRESS_NAMED("ark vm concat!");
                         break;
                     }