+++ src/arkscript/main.cpp
@@ -5,12 +5,10 @@
 #include <limits>
 
 #include <clipp.h>
-#define NOMINMAX
-#include <termcolor/termcolor.hpp>
+#include <termcolor/proxy.hpp>
 
 #include <Ark/Ark.hpp>
 #include <Ark/REPL/Repl.hpp>
-#include <Ark/Profiling.hpp>
 
 #include <Ark/Files.hpp>
 #include <Ark/Compiler/JsonCompiler.hpp>